fix redirectLocation action
When a redirectLocation is given, we need to make sure, it creates a new item on the history stack.
In the current implementation the ongoing transition, which is supposed to create the stack item, is stopped and the top item is replaced with the new location. This causes links not to occur in the history, if they trigger redirects.

There are two ways to fix this:

1) execute continueTransition() before the redirect (Pushing the original location onto the stack and replacing it afterwards)
2) cancel the ongoing transition and replace the action of the redirectLocation with "PUSH"

The second solution should probably be a bit faster.